+++ spec/keycloak-api-rails/helper_spec.rb
@@ -1,4 +1,65 @@
 RSpec.describe KeycloakApiRails::Helper do
+  describe "#read_token_from_query_string" do
+
+    subject { KeycloakApiRails::Helper.read_token_from_query_string(uri) }
+
+    context "when the uri is nil" do
+      let(:uri) { nil }
+      it "returns an empty token" do
+        expect(subject).to eq ""
+      end
+    end
+
+    context "when the uri is empty" do
+      let(:uri) { "" }
+      it "returns an empty token" do
+        expect(subject).to eq ""
+      end
+    end
+
+    context "when the uri only contains whitespaces" do
+      let(:uri) { "   " }
+      it "returns an empty token" do
+        expect(subject).to eq ""
+      end
+    end
+
+    context "when the uri has no query string" do
+      let(:uri) { "http://www.an-url.io/health" }
+      it "returns no token" do
+        expect(subject).to be_nil
+      end
+    end
+
+    context "when the uri has a query string but no token" do
+      let(:uri) { "http://www.an-url.io/health?firstName=ouioui" }
+      it "returns no token" do
+        expect(subject).to be_nil
+      end
+    end
+
+    context "when the uri has a token in the query string" do
+      let(:uri) { "http://www.an-url.io/health?authorizationToken=aToken" }
+      it "returns the token" do
+        expect(subject).to eq "aToken"
+      end
+    end
+
+    context "when the uri has a token among other query string parameters" do
+      let(:uri) { "http://www.an-url.io/health?firstName=ouioui&authorizationToken=aToken&lastName=nonnon" }
+      it "returns the token" do
+        expect(subject).to eq "aToken"
+      end
+    end
+
+    context "when the uri has an empty token in the query string" do
+      let(:uri) { "http://www.an-url.io/health?authorizationToken=" }
+      it "returns an empty token" do
+        expect(subject).to eq ""
+      end
+    end
+  end
+
   describe "#create_url_with_token" do
 
     let(:uri)   { "http://www.an-url.io" }